fixing bug when running pipeline with a subset of the default preprocessing steps
Running a pipeline with preprocessing kwarg "run_outlier_cleaning=False" fails. This is because the subsequent savgol filter looks for a variable called "position_processed". Without the outlier cleaning step, this variable is not created, the most recent output being "position_egocentric_aligned". This commit creates a flexible "latest_output" variable that is updated across the preprocessing steps and passed from one step to the next executed step. the preprocessing function returns this "latest_output" so that the pipeline can then pass it to "create_trainset"